tools: fix fbthrift

+++ tools/install_fbthrift.sh
@@ -1,4 +1,7 @@
 installCheck () {
+  if [ ! -d fbthrift ]; then
+    return 1
+  fi
   python -mthrift_compiler.main --gen cpp2 hello_world.thrift
   if [ -d gen-cpp2 ]; then
     rm -rf gen-cpp2
